+++ lib/bar_chart.dart
@@ -2,5 +2,7 @@ library ssun_chart;
 
 import 'package:flutter/material.dart';
 
+import 'utils/value_formatter.dart';
+
 part 'model/bar_data.dart';
 part 'ui/bar_chart.dart';