+++ lib/radar_chart.dart
@@ -4,6 +4,8 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 
+import 'utils/value_formatter.dart';
+
 part 'model/radar_data.dart';
 part 'ui/radar_chart.dart';
 part 'utils/utils.dart';