+++ lib/ui/bar_chart.dart
@@ -4,12 +4,14 @@ class BarChart extends StatefulWidget {
   const BarChart({
     this.bgColor = Colors.black12,
     this.yMaximum = 100,
+    this.yValueFormatter,
     required this.data,
     super.key
   });
 
   final Color bgColor;
   final double yMaximum;
+  final ValueFormatter? yValueFormatter;
   final BarData data;
 
   @override
@@ -23,6 +25,7 @@ class _BarChartState extends State<BarChart> {
       painter: _BarChartCustomPainter(
         bgColor: widget.bgColor,
         yMaximum: widget.yMaximum,
+        yValueFormatter: widget.yValueFormatter,
         data: widget.data,
       ),
     );
@@ -32,26 +35,33 @@ class _BarChartState extends State<BarChart> {
 class _BarChartCustomPainter extends CustomPainter {
   static const double xMargin = 10;
   static const double yMargin = 10;
+  static const double yLabelMargin = 2;
   static const double setMargin = 10;
   static const double barMargin = 3;
 
   _BarChartCustomPainter({
     required this.bgColor,
     required this.yMaximum,
+    required this.yValueFormatter,
     required this.data,
   });
 
   final Color bgColor;
   final double yMaximum;
+  final ValueFormatter? yValueFormatter;
   final BarData data;
 
   @override
   void paint(Canvas canvas, Size size) {
-    final anchor = Offset(xMargin, size.height - yMargin); // x축, y축 만나는 지점
     const yGuideLineCount = 5;
 
     canvas.clipRect(Rect.fromLTWH(0, 0, size.width, size.height));
     canvas.drawRect(Rect.fromLTWH(0, 0, size.width, size.height), Paint()..color = bgColor);
+
+    // y축 Label
+    final yLabelWidth = drawYLabels(canvas, size, yGuideLineCount);
+
+    final anchor = Offset(yLabelWidth, size.height - yMargin); // x축, y축 만나는 지점
 
     // 데이터
     drawData(canvas, size, anchor);
@@ -64,9 +74,6 @@ class _BarChartCustomPainter extends CustomPainter {
 
     // y축 GuideLine
     drawYGuideLines(canvas, size, anchor, yGuideLineCount);
-
-    // y축 Label
-    drawYLabels(canvas, size, anchor, yGuideLineCount);
   }
 
   void drawXAxis(Canvas canvas, Size size, Offset anchor) {
@@ -96,12 +103,13 @@ class _BarChartCustomPainter extends CustomPainter {
     }
   }
 
-  void drawYLabels(Canvas canvas, Size size, Offset anchor, int yGuideLineCount) {
+  double drawYLabels(Canvas canvas, Size size, int yGuideLineCount) {
+    double maxWidth = 0.0;
+    List<TextPainter> painters = List.empty(growable: true);
     for (int i = 0; i <= yGuideLineCount; i++) {
-      final y = anchor.dy - (size.height - yMargin * 2) / yGuideLineCount * i;
       final textPainter = TextPainter()
         ..text = TextSpan(
-          text: (yMaximum / yGuideLineCount * i).toString(),
+            text: yValueFormatter != null ? yValueFormatter!.format(yMaximum / yGuideLineCount * i) : (yMaximum / yGuideLineCount * i).toString(),
           style: const TextStyle(
             color: Colors.black,
             fontSize: 10
@@ -111,14 +119,25 @@ class _BarChartCustomPainter extends CustomPainter {
         ..textAlign = TextAlign.center
         ..layout();
 
+      if (textPainter.width > maxWidth) {
+        maxWidth = textPainter.width;
+      }
+
+      painters.add(textPainter);
+    }
+
+    for (int i = 0; i <= yGuideLineCount; i++) {
+      final y = (size.height - yMargin) - (size.height - yMargin * 2) / yGuideLineCount * i;
+      final textPainter = painters[i];
       textPainter.paint(
         canvas,
         Offset(
-          anchor.dx - textPainter.width / 2,
+          maxWidth - textPainter.width + yLabelMargin,
           y - textPainter.height / 2
         )
       );
     }
+    return maxWidth + yLabelMargin + yLabelMargin;
   }
 
   void drawData(Canvas canvas, Size size, Offset anchor) {