+++ lib/ui/line_chart.dart
@@ -6,6 +6,7 @@ class LineChart extends StatefulWidget {
     this.markerSize = 2.0,
     this.drawMarker = false,
     this.yMaximum = 100,
+    this.yValueFormatter,
     required this.data,
     super.key
   });
@@ -14,6 +15,7 @@ class LineChart extends StatefulWidget {
   final double markerSize;
   final bool drawMarker;
   final double yMaximum;
+  final ValueFormatter? yValueFormatter;
   final LineData data;
 
   @override
@@ -30,6 +32,7 @@ class _LineChartState extends State<LineChart> {
         markerSize: widget.markerSize,
         drawMarker: widget.drawMarker,
         yMaximum: widget.yMaximum,
+        yValueFormatter: widget.yValueFormatter,
         data: widget.data,
       ),
     );
@@ -39,12 +42,14 @@ class _LineChartState extends State<LineChart> {
 class _LineChartCustomPaint extends CustomPainter {
   static const double xMargin = 10;
   static const double yMargin = 10;
+  static const double yLabelMargin = 2;
 
   _LineChartCustomPaint({
     required this.bgColor,
     required this.markerSize,
     required this.drawMarker,
     required this.yMaximum,
+    required this.yValueFormatter,
     required this.data,
   });
 
@@ -52,15 +57,20 @@ class _LineChartCustomPaint extends CustomPainter {
   final double markerSize;
   final bool drawMarker;
   final double yMaximum;
+  final ValueFormatter? yValueFormatter;
   final LineData data;
 
   @override
   void paint(Canvas canvas, Size size) {
-    final anchor = Offset(xMargin, size.height - yMargin); // x축, y축 만나는 지점
     const yGuideLineCount = 5;
 
     canvas.clipRect(Rect.fromLTWH(0, 0, size.width, size.height));
     canvas.drawRect(Rect.fromLTWH(0, 0, size.width, size.height), Paint()..color = bgColor);
+
+    // y축 Label
+    final yLabelWidth = drawYLabels(canvas, size, yGuideLineCount);
+
+    final anchor = Offset(yLabelWidth, size.height - yMargin); // x축, y축 만나는 지점
 
     // 데이터
     drawData(canvas, size, anchor);
@@ -73,9 +83,6 @@ class _LineChartCustomPaint extends CustomPainter {
 
     // y축 GuideLine
     drawYGuideLines(canvas, size, anchor, yGuideLineCount);
-
-    // y축 Label
-    drawYLabels(canvas, size, anchor, yGuideLineCount);
   }
 
   void drawXAxis(Canvas canvas, Size size, Offset anchor) {
@@ -105,12 +112,13 @@ class _LineChartCustomPaint extends CustomPainter {
     }
   }
 
-  void drawYLabels(Canvas canvas, Size size, Offset anchor, int yGuideLineCount) {
+  double drawYLabels(Canvas canvas, Size size, int yGuideLineCount) {
+    double maxWidth = 0.0;
+    List<TextPainter> painters = List.empty(growable: true);
     for (int i = 0; i <= yGuideLineCount; i++) {
-      final y = anchor.dy - (size.height - yMargin * 2) / yGuideLineCount * i;
       final textPainter = TextPainter()
         ..text = TextSpan(
-          text: (yMaximum / yGuideLineCount * i).toString(),
+          text: yValueFormatter != null ? yValueFormatter!.format(yMaximum / yGuideLineCount * i) : (yMaximum / yGuideLineCount * i).toString(),
           style: const TextStyle(
             color: Colors.black,
             fontSize: 10
@@ -120,14 +128,25 @@ class _LineChartCustomPaint extends CustomPainter {
         ..textAlign = TextAlign.center
         ..layout();
 
+      if (textPainter.width > maxWidth) {
+        maxWidth = textPainter.width;
+      }
+
+      painters.add(textPainter);
+    }
+
+    for (int i = 0; i <= yGuideLineCount; i++) {
+      final y = (size.height - yMargin) - (size.height - yMargin * 2) / yGuideLineCount * i;
+      final textPainter = painters[i];
       textPainter.paint(
-          canvas,
-          Offset(
-            anchor.dx - textPainter.width / 2,
-            y - textPainter.height / 2
-          )
+        canvas,
+        Offset(
+          maxWidth - textPainter.width + yLabelMargin,
+          y - textPainter.height / 2
+        )
       );
     }
+    return maxWidth + yLabelMargin + yLabelMargin;
   }
 
   void drawData(Canvas canvas, Size size, Offset anchor) {
@@ -183,7 +202,7 @@ class _LineChartCustomPaint extends CustomPainter {
         }
         fillPath.lineTo(x2, y2);
         if (j == entry.entries.length - 2) {
-          fillPath.lineTo(size.width - xMargin, anchor.dy);
+          fillPath.lineTo(size.width - anchor.dx, anchor.dy);
           fillPath.lineTo(anchor.dx, anchor.dy);
         }
       }