+++ CHANGELOG.md
@@ -12,3 +12,15 @@
 
 ## 0.0.5
 * Add Pie Chart Detail
+
+## 0.0.6
+* Add Radar Chart Detail
+
+## 0.0.7
+* Line Chart, Bar Chart Y-Axis Label Detail
+
+## 0.0.8
+* Y-Axis Label Formatting
+
+## 0.0.9
+* Add README.md